Move services to ServiceContext static singleton class
